# server/src/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Donggle Swagger API!",
    "version": "1.0.0",
    "title": "Donggle",
    "contact": {
      "email": "notready@donggle"
    }
  },
  "host": "boost.boost-donggle.kro.kr",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/user/profile/image": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "유저의 Profile Image 업데이트",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Profile Image object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserProfileImage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "유저 프로필 업데이트 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "properties": {
                    "image": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "example": "이미지를 성공적으로 업데이트 했습니다."
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다!"
                }
              }
            }
          },
          "500": {
            "description": "유저 프로필 업데이트 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "내부 서버 오류입니다!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "유저의 Profile Image 삭제",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Profile Image object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserProfileImage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "유저 프로필 이미지 업데이트 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "성공적으로 이미지를 삭제했습니다."
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다!"
                }
              }
            }
          },
          "500": {
            "description": "유저 프로필 이미지 업데이트 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "내부 서버 오류입니다!"
                }
              }
            }
          }
        }
      }
    },
    "/user/profile/address": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "유저의 Profile address 업데이트",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Profile address object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserProfileAddress"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "유저 프로필 주소 업데이트 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "example": "주소 업데이트에 성공했습니다."
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다!"
                }
              }
            }
          },
          "500": {
            "description": "유저 프로필 주소 업데이트 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "내부 서버 오류입니다!"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "UserProfileImage": {
      "type": "object",
      "properties": {
        "oauth_email": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "UserProfileAddress": {
      "type": "object",
      "properties": {
        "oauth_email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      }
    }
  }
}